moving more tests (#6680)
