+++ conans/test/unittests/client/tools/win/vcvars_clangcl_test.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4
 
 import platform
 import unittest
